+++ src/rb.hpp
@@ -28,8 +28,8 @@ void deleteRB(RedBlack **T, RedBlack *valor);
 void removeRB(RedBlack **raiz,RedBlack **t, Record r);
 RedBlack * TreeMinimun(RedBlack **t);
 
-long long insercaoRB(RedBlack **Raiz,vector<double> numeros,int x);
-long long remocaoRB(RedBlack **Raiz,vector<double> busca);
+long insercaoRB(RedBlack **Raiz,vector<double> numeros,int x);
+long remocaoRB(RedBlack **Raiz,vector<double> busca);
 
 
 #endif
